improve db and migrations

+++ backend/XPos.Domain/Dtos/QuotationDtos.cs
@@ -5,9 +5,9 @@ public class CreateQuotationDto
     public DateTime Date { get; set; }
     public long ClientId { get; set; }
     public long WarehouseId { get; set; }
-    public double? TaxRate { get; set; }
-    public double? Discount { get; set; }
-    public double? Shipping { get; set; }
+    public decimal? TaxRate { get; set; }
+    public decimal? Discount { get; set; }
+    public decimal? Shipping { get; set; }
     public string Status { get; set; } = "pending";
     public string? Notes { get; set; }
     public List<CreateQuotationDetailDto> Details { get; set; } = new();
@@ -16,8 +16,8 @@ public class CreateQuotationDto
 public class CreateQuotationDetailDto
 {
     public long ProductId { get; set; }
-    public double Price { get; set; }
-    public double Quantity { get; set; }
+    public decimal Price { get; set; }
+    public decimal Quantity { get; set; }
 }
 
 public class UpdateQuotationDto : CreateQuotationDto
