+++ backend/XPos.Domain/Dtos/ReturnDtos.cs
@@ -5,8 +5,8 @@ public class CreateSaleReturnDto
     public DateTime Date { get; set; }
     public long ClientId { get; set; }
     public long WarehouseId { get; set; }
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = "completed";
     public string PaymentStatus { get; set; } = "unpaid";
     public List<CreateSaleReturnDetailDto> Details { get; set; } = new();
@@ -15,8 +15,8 @@ public class CreateSaleReturnDto
 public class CreateSaleReturnDetailDto
 {
     public long ProductId { get; set; }
-    public double Price { get; set; }
-    public double Quantity { get; set; }
+    public decimal Price { get; set; }
+    public decimal Quantity { get; set; }
 }
 
 public class CreatePurchaseReturnDto
@@ -24,8 +24,8 @@ public class CreatePurchaseReturnDto
     public DateTime Date { get; set; }
     public long ProviderId { get; set; }
     public long WarehouseId { get; set; }
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = "completed";
     public string PaymentStatus { get; set; } = "unpaid";
     public List<CreatePurchaseReturnDetailDto> Details { get; set; } = new();
@@ -34,6 +34,6 @@ public class CreatePurchaseReturnDto
 public class CreatePurchaseReturnDetailDto
 {
     public long ProductId { get; set; }
-    public double Cost { get; set; }
-    public double Quantity { get; set; }
+    public decimal Cost { get; set; }
+    public decimal Quantity { get; set; }
 }
