+++ backend/XPos.Domain/Dtos/TransferDtos.cs
@@ -5,9 +5,9 @@ public class CreateTransferDto
     public DateTime Date { get; set; }
     public long FromWarehouseId { get; set; }
     public long ToWarehouseId { get; set; }
-    public double TaxRate { get; set; }
-    public double Discount { get; set; }
-    public double Shipping { get; set; }
+    public decimal TaxRate { get; set; }
+    public decimal Discount { get; set; }
+    public decimal Shipping { get; set; }
     public string Status { get; set; } = "Completed";
     public string? Notes { get; set; }
     public List<CreateTransferDetailDto> Details { get; set; } = new();
@@ -16,8 +16,8 @@ public class CreateTransferDto
 public class CreateTransferDetailDto
 {
     public long ProductId { get; set; }
-    public double Cost { get; set; }
-    public double Quantity { get; set; }
+    public decimal Cost { get; set; }
+    public decimal Quantity { get; set; }
 }
 
 public class UpdateTransferDto : CreateTransferDto
