+++ backend/XPos.Domain/Interfaces/IInventoryRepository.cs
@@ -4,5 +4,5 @@ namespace XPos.Domain.Interfaces;
 
 public interface IInventoryRepository
 {
-    Task UpdateStockAsync(long productId, long warehouseId, double quantityChange);
+    Task UpdateStockAsync(long productId, long warehouseId, decimal quantityChange);
 }
