+++ backend/XPos.Domain/Models/ReportModels.cs
@@ -6,8 +6,8 @@ public class SalesReportDto
     public string Ref { get; set; } = string.Empty;
     public string ClientName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string PaymentStatus { get; set; } = string.Empty;
 }
 
@@ -17,8 +17,8 @@ public class PurchaseReportDto
     public string Ref { get; set; } = string.Empty;
     public string ProviderName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = string.Empty;
 }
 
@@ -28,33 +28,33 @@ public class StockReportDto
     public string ProductName { get; set; } = string.Empty;
     public string CategoryName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double Quantity { get; set; }
-    public double StockAlert { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal StockAlert { get; set; }
 }
 
 public class ProfitLossReportDto
 {
-    public double TotalSales { get; set; }
-    public double TotalPurchases { get; set; }
-    public double TotalExpenses { get; set; }
-    public double TotalReturns { get; set; }
-    public double NetProfit => TotalSales - TotalPurchases - TotalExpenses - TotalReturns;
+    public decimal TotalSales { get; set; }
+    public decimal TotalPurchases { get; set; }
+    public decimal TotalExpenses { get; set; }
+    public decimal TotalReturns { get; set; }
+    public decimal NetProfit => TotalSales - TotalPurchases - TotalExpenses - TotalReturns;
 }
 
 public class RecentSaleDto
 {
     public string Ref { get; set; } = string.Empty;
     public string ClientName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
+    public decimal GrandTotal { get; set; }
     public string Status { get; set; } = string.Empty;
 }
 
 public class DashboardSummaryDto
 {
-    public double TodaySales { get; set; }
+    public decimal TodaySales { get; set; }
     public int TodaySalesCount { get; set; }
-    public double MonthlySales { get; set; }
-    public double MonthlyPurchases { get; set; }
+    public decimal MonthlySales { get; set; }
+    public decimal MonthlyPurchases { get; set; }
     public List<RecentSaleDto> RecentSales { get; set; } = new();
     public List<TopProductDto> TopProducts { get; set; } = new();
 }
@@ -62,8 +62,8 @@ public class DashboardSummaryDto
 public class TopProductDto
 {
     public string Name { get; set; } = string.Empty;
-    public double Quantity { get; set; }
-    public double Total { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal Total { get; set; }
 }
 
 public class ProductMovementReportDto
@@ -71,7 +71,7 @@ public class ProductMovementReportDto
     public DateTime Date { get; set; }
     public string Ref { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty; // 'Sale', 'Purchase', 'Transfer', 'Adjustment'
-    public double Quantity { get; set; }
+    public decimal Quantity { get; set; }
     public string WarehouseName { get; set; } = string.Empty;
 }
 
@@ -79,8 +79,8 @@ public class StockAlertReportDto
 {
     public string Code { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
-    public double Quantity { get; set; }
-    public double StockAlert { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal StockAlert { get; set; }
 }
 
 public class ActivityReportDto
@@ -88,7 +88,7 @@ public class ActivityReportDto
     public DateTime Date { get; set; }
     public string Ref { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
-    public double Total { get; set; }
+    public decimal Total { get; set; }
     public string WarehouseName { get; set; } = string.Empty;
     public string UserName { get; set; } = string.Empty;
 }
@@ -101,9 +101,9 @@ public class ClientReportDto
     public string Phone { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public int TotalSales { get; set; }
-    public double TotalAmount { get; set; }
-    public double TotalPaid { get; set; }
-    public double DueAmount => TotalAmount - TotalPaid; // Deuda pendiente
+    public decimal TotalAmount { get; set; }
+    public decimal TotalPaid { get; set; }
+    public decimal DueAmount => TotalAmount - TotalPaid; // Deuda pendiente
 }
 
 public class ProviderReportDto
@@ -113,7 +113,7 @@ public class ProviderReportDto
     public string Phone { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
     public int TotalPurchases { get; set; }
-    public double TotalAmount { get; set; }
-    public double TotalPaid { get; set; }
-    public double DueAmount => TotalAmount - TotalPaid; // Deuda pendiente
+    public decimal TotalAmount { get; set; }
+    public decimal TotalPaid { get; set; }
+    public decimal DueAmount => TotalAmount - TotalPaid; // Deuda pendiente
 }
