+++ backend/XPos.Domain/Models/ReturnModels.cs
@@ -9,12 +9,12 @@ public class SaleReturn
     public long? SaleId { get; set; }
     public long ClientId { get; set; }
     public long WarehouseId { get; set; }
-    public double? TaxRate { get; set; }
-    public double? TaxNet { get; set; }
-    public double? Discount { get; set; }
-    public double? Shipping { get; set; }
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal? TaxRate { get; set; }
+    public decimal? TaxNet { get; set; }
+    public decimal? Discount { get; set; }
+    public decimal? Shipping { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = "completed";
     public string PaymentStatus { get; set; } = "unpaid";
     public string? Notes { get; set; }
@@ -39,13 +39,13 @@ public class SaleReturnDetail
     public long ProductId { get; set; }
     public long? ProductVariantId { get; set; }
     public long? SaleUnitId { get; set; }
-    public double Price { get; set; }
-    public double? TaxNet { get; set; }
+    public decimal Price { get; set; }
+    public decimal? TaxNet { get; set; }
     public string? TaxMethod { get; set; }
-    public double? Discount { get; set; }
+    public decimal? Discount { get; set; }
     public string? DiscountMethod { get; set; }
-    public double Total { get; set; }
-    public double Quantity { get; set; }
+    public decimal Total { get; set; }
+    public decimal Quantity { get; set; }
     
     public DateTime? CreatedAt { get; set; }
     public DateTime? UpdatedAt { get; set; }
@@ -61,12 +61,12 @@ public class PurchaseReturn
     public long? PurchaseId { get; set; }
     public long ProviderId { get; set; }
     public long WarehouseId { get; set; }
-    public double? TaxRate { get; set; }
-    public double? TaxNet { get; set; }
-    public double? Discount { get; set; }
-    public double? Shipping { get; set; }
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal? TaxRate { get; set; }
+    public decimal? TaxNet { get; set; }
+    public decimal? Discount { get; set; }
+    public decimal? Shipping { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = "completed";
     public string PaymentStatus { get; set; } = "unpaid";
     public string? Notes { get; set; }
@@ -91,13 +91,13 @@ public class PurchaseReturnDetail
     public long ProductId { get; set; }
     public long? ProductVariantId { get; set; }
     public long? PurchaseUnitId { get; set; }
-    public double Cost { get; set; }
-    public double? TaxNet { get; set; }
+    public decimal Cost { get; set; }
+    public decimal? TaxNet { get; set; }
     public string? TaxMethod { get; set; }
-    public double? Discount { get; set; }
+    public decimal? Discount { get; set; }
     public string? DiscountMethod { get; set; }
-    public double Total { get; set; }
-    public double Quantity { get; set; }
+    public decimal Total { get; set; }
+    public decimal Quantity { get; set; }
     
     public DateTime? CreatedAt { get; set; }
     public DateTime? UpdatedAt { get; set; }
@@ -111,7 +111,7 @@ public class SaleReturnReadDto
     public DateTime Date { get; set; }
     public string ClientName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
+    public decimal GrandTotal { get; set; }
     public string Status { get; set; } = string.Empty;
     public string PaymentStatus { get; set; } = string.Empty;
     public long? VoucherId { get; set; }
@@ -124,7 +124,7 @@ public class PurchaseReturnReadDto
     public DateTime Date { get; set; }
     public string ProviderName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
+    public decimal GrandTotal { get; set; }
     public string Status { get; set; } = string.Empty;
     public string PaymentStatus { get; set; } = string.Empty;
     public long? VoucherId { get; set; }
