+++ backend/XPos.Domain/Models/SaleModels.cs
@@ -9,12 +9,12 @@ public class Sale
     public bool IsPos { get; set; }
     public long ClientId { get; set; }
     public long WarehouseId { get; set; }
-    public double? TaxRate { get; set; }
-    public double? TaxNet { get; set; }
-    public double? Discount { get; set; }
-    public double? Shipping { get; set; }
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal? TaxRate { get; set; }
+    public decimal? TaxNet { get; set; }
+    public decimal? Discount { get; set; }
+    public decimal? Shipping { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = "completed";
     public string PaymentStatus { get; set; } = "unpaid";
     public string? ShippingStatus { get; set; }
@@ -41,13 +41,13 @@ public class SaleDetail
     public long ProductId { get; set; }
     public long? ProductVariantId { get; set; }
     public long? SaleUnitId { get; set; }
-    public double Price { get; set; }
-    public double? TaxNet { get; set; }
+    public decimal Price { get; set; }
+    public decimal? TaxNet { get; set; }
     public string? TaxMethod { get; set; }
-    public double? Discount { get; set; }
+    public decimal? Discount { get; set; }
     public string? DiscountMethod { get; set; }
-    public double Total { get; set; }
-    public double Quantity { get; set; }
+    public decimal Total { get; set; }
+    public decimal Quantity { get; set; }
     
     // Audit fields
     public DateTime? CreatedAt { get; set; }
@@ -66,8 +66,8 @@ public class SaleReadDto
     public DateTime Date { get; set; }
     public string ClientName { get; set; } = string.Empty;
     public string WarehouseName { get; set; } = string.Empty;
-    public double GrandTotal { get; set; }
-    public double PaidAmount { get; set; }
+    public decimal GrandTotal { get; set; }
+    public decimal PaidAmount { get; set; }
     public string Status { get; set; } = string.Empty;
     public string PaymentStatus { get; set; } = string.Empty;
     public long? VoucherId { get; set; }
