+++ backend/XPos.Domain/Models/TransferModels.cs
@@ -8,12 +8,12 @@ public class Transfer
     public DateTime Date { get; set; }
     public long FromWarehouseId { get; set; }
     public long ToWarehouseId { get; set; }
-    public double Items { get; set; }
-    public double TaxRate { get; set; }
-    public double TaxNet { get; set; }
-    public double Discount { get; set; }
-    public double Shipping { get; set; }
-    public double GrandTotal { get; set; }
+    public decimal Items { get; set; }
+    public decimal TaxRate { get; set; }
+    public decimal TaxNet { get; set; }
+    public decimal Discount { get; set; }
+    public decimal Shipping { get; set; }
+    public decimal GrandTotal { get; set; }
     public string Status { get; set; } = "Completed";
     public string? Notes { get; set; }
     
@@ -35,13 +35,13 @@ public class TransferDetail
     public long ProductId { get; set; }
     public long? ProductVariantId { get; set; }
     public long? PurchaseUnitId { get; set; }
-    public double Cost { get; set; }
-    public double TaxNet { get; set; }
+    public decimal Cost { get; set; }
+    public decimal TaxNet { get; set; }
     public string TaxMethod { get; set; } = "1";
-    public double Discount { get; set; }
+    public decimal Discount { get; set; }
     public string DiscountMethod { get; set; } = "1";
-    public double Quantity { get; set; }
-    public double Total { get; set; }
+    public decimal Quantity { get; set; }
+    public decimal Total { get; set; }
 }
 
 public class TransferReadDto
@@ -51,7 +51,7 @@ public class TransferReadDto
     public DateTime Date { get; set; }
     public string FromWarehouseName { get; set; } = string.Empty;
     public string ToWarehouseName { get; set; } = string.Empty;
-    public double Items { get; set; }
-    public double GrandTotal { get; set; }
+    public decimal Items { get; set; }
+    public decimal GrandTotal { get; set; }
     public string Status { get; set; } = string.Empty;
 }
